backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
channels>=4.0
daphne>=4.0
mediapipe>=0.10
tensorflow>=2.13
opencv-python>=4.8
numpy>=1.24
pyttsx3>=2.90
pywhatkit>=5.4
Pillow>=10.0
scikit-learn>=1.3
matplotlib>=3.7
